demo/requirements.txt
Chameleon==2.9.2
Mako==0.7.1
MarkupSafe==0.15
PasteDeploy==1.5.0
WebOb==1.2
pyramid==1.3.2
-e git+git@github.com:blaflamme/pyramid_dotcloud
repoze.lru==0.5
translationstring==1.1
venusian==1.0a6
waitress==0.8.1
wsgiref==0.1.2
zope.deprecation==4.0.0
zope.interface==4.0.1
